Move dependency cache to a separate module
