+++ language/tools/move-package/src/lib.rs
@@ -13,11 +13,14 @@ use anyhow::{bail, Result};
 use clap::*;
 use move_core_types::account_address::AccountAddress;
 use move_model::model::GlobalEnv;
-use resolution::{dependency_graph::DependencyGraph, resolution_graph::ResolvedGraph};
+use resolution::{
+    dependency_cache::DependencyCache, dependency_graph::DependencyGraph,
+    resolution_graph::ResolvedGraph,
+};
 use serde::{Deserialize, Serialize};
 use source_package::layout::SourcePackageLayout;
 use std::{
-    collections::{BTreeMap, BTreeSet},
+    collections::BTreeMap,
     fmt,
     io::Write,
     path::{Path, PathBuf},
@@ -226,22 +229,16 @@ impl BuildConfig {
         // possibly be set by a different process in parallel.
         let manifest = manifest_parser::parse_source_manifest(toml_manifest)?;
 
+        let mut dependency_cache = DependencyCache::new(self.skip_fetch_latest_git_deps);
         let dependency_graph =
-            DependencyGraph::new(&manifest, path, self.skip_fetch_latest_git_deps, writer)?;
-        let mut fetched_deps = BTreeSet::new();
+            DependencyGraph::new(&manifest, path, &mut dependency_cache, writer)?;
 
         let lock = dependency_graph.write_to_lock()?;
         if let Some(lock_path) = &self.lock_file {
             lock.commit(lock_path)?;
         }
 
-        let mut resolved_graph =
-            ResolvedGraph::resolve(dependency_graph, self, writer, &mut fetched_deps)?;
-        // re-insert information about dependencies fetched during resolution (use a separate set
-        // structure to avoid passing the entire dependency graph as mutable to the resolve
-        // function)
-        resolved_graph.graph.fetched_deps = fetched_deps;
-        Ok(resolved_graph)
+        ResolvedGraph::resolve(dependency_graph, self, &mut dependency_cache, writer)
     }
 
     fn parse_toml_manifest(&self, path: PathBuf) -> Result<toml::Value> {
