+++ language/tools/move-package/tests/test_additional_addresses.rs
@@ -4,14 +4,13 @@
 
 use move_core_types::account_address::AccountAddress;
 use move_package::{
-    resolution::{dependency_graph as DG, resolution_graph as RG},
+    resolution::{
+        dependency_cache::DependencyCache, dependency_graph as DG, resolution_graph as RG,
+    },
     source_package::manifest_parser as MP,
     BuildConfig,
 };
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    path::PathBuf,
-};
+use std::{collections::BTreeMap, path::PathBuf};
 use tempfile::tempdir;
 
 #[test]
@@ -26,13 +25,9 @@ fn test_additonal_addresses() {
 
     let pm = MP::parse_move_manifest_from_file(&path).unwrap();
 
+    let mut dependency_cache = DependencyCache::new(/* skip_fetch_latest_git_deps */ true);
     let mut sink = std::io::sink();
-    let dg = DG::DependencyGraph::new(
-        &pm, path, /* skip_fetch_latest_git_deps */ true, &mut sink,
-    )
-    .unwrap();
-
-    let mut fetched_deps = BTreeSet::new();
+    let dg = DG::DependencyGraph::new(&pm, path, &mut dependency_cache, &mut sink).unwrap();
 
     assert!(RG::ResolvedGraph::resolve(
         dg.clone(),
@@ -44,8 +39,8 @@ fn test_additonal_addresses() {
             )]),
             ..Default::default()
         },
+        &mut dependency_cache,
         &mut sink,
-        &mut fetched_deps.clone(),
     )
     .is_ok());
 
@@ -55,8 +50,8 @@ fn test_additonal_addresses() {
             install_dir: Some(tempdir().unwrap().path().to_path_buf()),
             ..Default::default()
         },
+        &mut dependency_cache,
         &mut sink,
-        &mut fetched_deps,
     )
     .is_err());
 }
@@ -69,13 +64,9 @@ fn test_additonal_addresses_already_assigned_same_value() {
 
     let pm = MP::parse_move_manifest_from_file(&path).unwrap();
 
+    let mut dependency_cache = DependencyCache::new(/* skip_fetch_latest_git_deps */ true);
     let mut sink = std::io::sink();
-    let dg = DG::DependencyGraph::new(
-        &pm, path, /* skip_fetch_latest_git_deps */ true, &mut sink,
-    )
-    .unwrap();
-
-    let mut fetched_deps = BTreeSet::new();
+    let dg = DG::DependencyGraph::new(&pm, path, &mut dependency_cache, &mut sink).unwrap();
 
     assert!(RG::ResolvedGraph::resolve(
         dg,
@@ -87,8 +78,8 @@ fn test_additonal_addresses_already_assigned_same_value() {
             )]),
             ..Default::default()
         },
+        &mut dependency_cache,
         &mut sink,
-        &mut fetched_deps,
     )
     .is_ok());
 }
@@ -101,13 +92,9 @@ fn test_additonal_addresses_already_assigned_different_value() {
 
     let pm = MP::parse_move_manifest_from_file(&path).unwrap();
 
+    let mut dependency_cache = DependencyCache::new(/* skip_fetch_latest_git_deps */ true);
     let mut sink = std::io::sink();
-    let dg = DG::DependencyGraph::new(
-        &pm, path, /* skip_fetch_latest_git_deps */ true, &mut sink,
-    )
-    .unwrap();
-
-    let mut fetched_deps = BTreeSet::new();
+    let dg = DG::DependencyGraph::new(&pm, path, &mut dependency_cache, &mut sink).unwrap();
 
     assert!(RG::ResolvedGraph::resolve(
         dg,
@@ -119,8 +106,8 @@ fn test_additonal_addresses_already_assigned_different_value() {
             )]),
             ..Default::default()
         },
+        &mut dependency_cache,
         &mut sink,
-        &mut fetched_deps,
     )
     .is_err());
 }
