+++ language/tools/move-package/tests/test_sources/basic_no_deps_address_not_assigned_with_dev_assignment/Move.resolved
@@ -9,7 +9,6 @@ ResolvedGraph {
         always_deps: {
             "test",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
