+++ language/tools/move-package/tests/test_sources/dep_dev_dep_diamond/Move.resolved
@@ -98,7 +98,6 @@ ResolvedGraph {
             "C",
             "Root",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
