+++ language/tools/move-package/tests/test_sources/dep_good_digest/Move.resolved
@@ -29,7 +29,6 @@ ResolvedGraph {
             "OtherDep",
             "Root",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
