+++ language/tools/move-package/tests/test_sources/diamond_problem_backflow_resolution/Move.resolved
@@ -73,7 +73,6 @@ ResolvedGraph {
             "C",
             "Root",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
