+++ language/tools/move-package/tests/test_sources/external_dev_dep/Move.resolved
@@ -89,7 +89,6 @@ ResolvedGraph {
             "ADep",
             "Root",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
