+++ language/tools/move-package/tests/test_sources/multiple_deps_rename/Move.resolved
@@ -47,7 +47,6 @@ ResolvedGraph {
             "D",
             "test",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
