+++ language/tools/move-package/tests/test_sources/one_dep_unification_across_local_renamings/Move.resolved
@@ -29,7 +29,6 @@ ResolvedGraph {
             "OtherDep",
             "Root",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
