+++ language/tools/move-package/tests/test_sources/parsing_invalid_identifier_package_name/Move.resolved
@@ -9,7 +9,6 @@ ResolvedGraph {
         always_deps: {
             "®´∑œ",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
