+++ language/tools/move-package/tests/test_sources/parsing_minimal_manifest/Move.resolved
@@ -9,7 +9,6 @@ ResolvedGraph {
         always_deps: {
             "name",
         },
-        fetched_deps: {},
     },
     build_options: BuildConfig {
         dev_mode: true,
